Update add_product.php
Updated with radio buttons to set the availability status of each product. Checking off "Yes" sets the availability to "1" (true), and checking off "No" sets the availability to "0" (false).
Wrote new way to pass the values into the query due to confusion with the passed values string syntax. Can be edited back into a string if desired.